Move Viewport and RTree to directories
