+++ packages/render/src/canvas.ts
@@ -1,4 +1,4 @@
-import Viewport from './Viewport.js'
+import Viewport from './viewport/Viewport.js'
 import CanvasRenderer from './renderers/CanvasRenderer.js'
 
 export { Viewport, CanvasRenderer }
