+++ packages/render/src/int-array.ts
@@ -1,4 +1,4 @@
-import Viewport from './Viewport.js'
+import Viewport from './viewport/Viewport.js'
 import IntArrayRenderer from './renderers/IntArrayRenderer.js'
 
 export { Viewport, IntArrayRenderer }
