+++ packages/render/src/shared/render-to-int-array.ts
@@ -1,7 +1,7 @@
 import classifyPoint from 'robust-point-in-polygon'
 
 import type WarpedMapList from '../maps/WarpedMapList.js'
-import type Viewport from '../Viewport.js'
+import type Viewport from '../viewport/Viewport.js'
 import type TileCache from '../tilecache/TileCache.js'
 import type { CachedTile } from '../tilecache/CacheableTile.js'
