+++ packages/render/src/webgl2.ts
@@ -1,4 +1,4 @@
-import Viewport from './Viewport.js'
+import Viewport from './viewport/Viewport.js'
 import WebGL2Renderer from './renderers/WebGL2Renderer.js'
 
 export { Viewport, WebGL2Renderer }
